Renamed appslist package to apps. When creating RequestsFragment and AppsDetailFragment we now pass all parameters as arguments using a bundle
