+++ twilio-authenticator-sample/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
         </activity>
 
         <activity
-            android:name=".appslist.AppsActivity"
+            android:name=".apps.AppsActivity"
             android:label="@string/title_activity_apps_list"
             android:theme="@style/AppTheme.NoActionBar"
             android:screenOrientation="portrait">
@@ -31,21 +31,21 @@
         <activity
             android:name=".approvalrequests.detail.ApprovalRequestDetailActivity"
             android:label="@string/title_activity_approval_request_detail"
-            android:parentActivityName=".appslist.MainActivity"
+            android:parentActivityName=".apps.MainActivity"
             android:screenOrientation="portrait">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".appslist.AppsActivity" />
+                android:value=".apps.AppsActivity" />
         </activity>
 
         <activity
-            android:name=".appslist.MainActivity"
+            android:name=".apps.MainActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".appslist.AppsActivity" />
+                android:value=".apps.AppsActivity" />
         </activity>
