+++ twilio-authenticator-sample/src/main/java/com/twilio/authenticatorsample/appdetail/AppDetailFragment.java
@@ -1,6 +1,5 @@
 package com.twilio.authenticatorsample.appdetail;
 
-import android.app.ActionBar;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.design.widget.Snackbar;
@@ -15,6 +14,8 @@ import com.twilio.authenticator.TwilioAuthenticator;
 import com.twilio.authenticator.external.App;
 import com.twilio.authenticator.external.AuthenticatorObserver;
 import com.twilio.authenticatorsample.R;
+import com.twilio.authenticatorsample.SampleApp;
+import com.twilio.authenticatorsample.apps.AppsActivity;
 import com.twilio.authenticatorsample.ui.views.AuthyTimerView;
 import com.twilio.authenticatorsample.utils.MessageHelper;
 
@@ -46,17 +47,17 @@ public class AppDetailFragment extends Fragment implements TokenTimer.OnTimerLis
      *
      * @return A new instance of fragment RequestsFragment.
      */
-    public static AppDetailFragment newInstance(Long appId, TwilioAuthenticator twilioAuthenticator) {
-
+    public static AppDetailFragment newInstance(Long appId) {
         AppDetailFragment appDetailFragment = new AppDetailFragment();
-        appDetailFragment.twilioAuthenticator = twilioAuthenticator;
-        appDetailFragment.appId = appId;
+        Bundle args = new Bundle();
+        args.putLong(AppsActivity.EXTRA_APP_ID, appId);
+        appDetailFragment.setArguments(args);
         return appDetailFragment;
 
     }
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
 
         View rootView = inflater.inflate(R.layout.fragment_app_detail, container, false);
         initViews(rootView);
@@ -89,6 +90,8 @@ public class AppDetailFragment extends Fragment implements TokenTimer.OnTimerLis
 
 
     private void initVars() {
+        appId = getArguments().getLong(AppsActivity.EXTRA_APP_ID);
+        twilioAuthenticator = ((SampleApp) getActivity().getApplicationContext()).getTwilioAuthenticator();
         messageHelper = new MessageHelper();
     }
