+++ twilio-authenticator-sample/src/main/res/layout/content_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.twilio.authenticatorsample.appslist.MainActivity"
+    tools:context="com.twilio.authenticatorsample.apps.MainActivity"
     tools:showIn="@layout/app_bar_main">
 
     <FrameLayout
